+++ lib/v1.0/C/pprz_transport.c
@@ -108,7 +108,7 @@ static void count_bytes(struct pprz_transport *trans __attribute__((unused)), st
 }
 
 static int check_available_space(struct pprz_transport *trans __attribute__((unused)), struct link_device *dev,
-                                 long *fd, uint8_t bytes)
+                                 long *fd, uint16_t bytes)
 {
   return dev->check_free_space(dev->periph, fd, bytes);
 }
